refactor: extract API types
